+++ packages/manager/modules/core/rollup.config.js
@@ -12,6 +12,8 @@ if (process.env.BUILD === 'production') {
     config.umd({
       output: {
         globals: {
+          '@ovh-ux/ng-ovh-http': 'ngOvhHttp',
+          '@ovh-ux/ng-ovh-sso-auth': 'ngOvhSsoAuth',
           angular: 'angular',
         },
       },