+++ packages/manager/modules/core/src/index.js
@@ -6,9 +6,9 @@ import 'angular-sanitize';
 import 'angular-resource';
 
 import kebabCase from 'lodash/kebabCase';
-import translateAsyncLoader from '@ovh-ux/translate-async-loader';
-import OvhHttp from 'ovh-angular-http';
-import ssoAuth from 'ovh-angular-sso-auth';
+import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
+import ngOvhHttp from '@ovh-ux/ng-ovh-http';
+import ngOvhSsoAuth from '@ovh-ux/ng-ovh-sso-auth';
 import ovhOuiAngularTranslations from './translate/ovh-ui-angular';
 
 import 'angular-dynamic-locale';
@@ -34,10 +34,10 @@ angular
     'ovh-api-services',
     'pascalprecht.translate',
     'tmh.dynamicLocale',
-    translateAsyncLoader,
+    ngTranslateAsyncLoader,
     ovhOuiAngularTranslations,
-    OvhHttp,
-    ssoAuth,
+    ngOvhHttp,
+    ngOvhSsoAuth,
   ])
   .constant('constants', {})
   .constant('CORE_LANGUAGES', LANGUAGES)
@@ -184,7 +184,7 @@ angular
     ]);
 
     $httpProvider.interceptors.push('serviceTypeInterceptor');
-    $httpProvider.interceptors.push('ssoAuthInterceptor');
+    $httpProvider.interceptors.push('OvhSsoAuthInterceptor');
   })
   .config((OvhHttpProvider) => {
     // OvhHttpProvider.rootPath = constants.swsProxyPath;