+++ packages/manager/modules/sms/src/options/telecom-sms-options.controller.js
@@ -7,15 +7,15 @@ export default class {
   $onInit() {
     this.actions = [{
       name: 'options_manage',
-      sref: 'sms.options.manage',
+      sref: 'sms.service.options.manage',
       text: this.$translate.instant('sms_options_manage'),
     }, {
       name: 'options_response',
-      sref: 'sms.options.response',
+      sref: 'sms.service.options.response',
       text: this.$translate.instant('sms_options_response'),
     }, {
       name: 'options_recredit',
-      sref: 'sms.options.recredit',
+      sref: 'sms.service.options.recredit',
       text: this.$translate.instant('sms_options_recredit'),
     }];
   }