+++ packages/manager/modules/sms/src/phonebooks/telecom-sms-phonebooks.component.js
@@ -1,5 +1,6 @@
 import angular from 'angular';
 import '@uirouter/angularjs';
+import 'angular-ui-validate';
 
 import controller from './telecom-sms-phonebooks.controller';
 import template from './telecom-sms-phonebooks.html';
@@ -11,13 +12,14 @@ const moduleName = 'ovhManagerSmsPhonebooksComponent';
 angular
   .module(moduleName, [
     'ui.router',
+    'ui.validate',
     create,
   ])
   .config(($stateProvider) => {
-    $stateProvider.state('sms.phonebooks', {
+    $stateProvider.state('sms.service.phonebooks', {
       url: '/phonebooks',
       views: {
-        'smsInnerView@sms': {
+        smsInnerView: {
           template,
           controller,
           controllerAs: 'PhonebooksCtrl',