+++ packages/manager/modules/sms/src/senders/add/index.js
@@ -5,10 +5,10 @@ import template from './telecom-sms-senders-add.html';
 const moduleName = 'ovhManagerSmsSendersAdd';
 
 angular.module(moduleName, []).config(($stateProvider) => {
-  $stateProvider.state('sms.senders.add', {
+  $stateProvider.state('sms.service.senders.add', {
     url: '/add',
     views: {
-      'smsView@sms': {
+      'smsView@sms.service': {
         template,
         controller,
         controllerAs: 'SmsSendersAddCtrl',