+++ packages/manager/modules/sms/src/sms/compose/index.js
@@ -7,10 +7,10 @@ import './addPhonebookContact/telecom-sms-sms-compose-addPhonebookContact.less';
 const moduleName = 'ovhManagerSmsSmsCompose';
 
 angular.module(moduleName, []).config(($stateProvider) => {
-  $stateProvider.state('sms.sms.compose', {
+  $stateProvider.state('sms.service.sms.compose', {
     url: '/compose',
     views: {
-      'smsView@sms': {
+      'smsView@sms.service': {
         template,
         controller,
         controllerAs: 'SmsComposeCtrl',