+++ packages/manager/modules/sms/src/sms/hlr/index.js
@@ -5,10 +5,10 @@ import template from './telecom-sms-sms-hlr.html';
 const moduleName = 'ovhManagerSmsSmsHlr';
 
 angular.module(moduleName, []).config(($stateProvider) => {
-  $stateProvider.state('sms.sms.hlr', {
+  $stateProvider.state('sms.service.sms.hlr', {
     url: '/hlr',
     views: {
-      'smsView@sms': {
+      'smsView@sms.service': {
         template,
         controller,
         controllerAs: 'SmsHlrCtrl',