+++ packages/manager/modules/sms/src/sms/outgoing/index.js
@@ -6,10 +6,10 @@ import template from './telecom-sms-sms-outgoing.html';
 const moduleName = 'ovhManagerSmsSmsOutgoing';
 
 angular.module(moduleName, []).config(($stateProvider) => {
-  $stateProvider.state('sms.sms.outgoing', {
+  $stateProvider.state('sms.service.sms.outgoing', {
     url: '/outgoing',
     views: {
-      'smsView@sms': {
+      'smsView@sms.service': {
         template,
         controller,
         controllerAs: 'SmsOutgoingCtrl',