+++ packages/manager/modules/sms/src/sms/templates/index.js
@@ -6,10 +6,10 @@ import template from './telecom-sms-sms-templates.html';
 const moduleName = 'ovhManagerSmsSmsTemplates';
 
 angular.module(moduleName, []).config(($stateProvider) => {
-  $stateProvider.state('sms.sms.templates', {
+  $stateProvider.state('sms.service.sms.templates', {
     url: '/templates',
     views: {
-      'smsView@sms': {
+      'smsView@sms.service': {
         template,
         controller,
         controllerAs: '$ctrl',