+++ packages/manager/modules/sms/src/telecom-sms.component.js
@@ -1,10 +1,12 @@
 import angular from 'angular';
 
-import '@ovh-ux/telecom-universe-components';
+import '@ovh-ux/ng-ovh-telecom-universe-components';
 import '@ovh-ux/manager-telecom-styles';
+import 'angular-messages';
 import 'ovh-angular-checkbox-table';
 import 'ovh-angular-responsive-tabs';
 import 'ovh-angular-pagination-front';
+import 'ovh-api-services';
 
 import 'ovh-ui-kit/dist/oui.css';
 import 'ovh-ui-kit-bs/dist/ovh-ui-kit-bs.css';
@@ -33,10 +35,12 @@ const moduleName = 'ovhManagerSmsComponent';
 
 angular
   .module(moduleName, [
-    'telecomUniverseComponents',
+    'ngOvhTelecomUniverseComponents',
     'ovh-angular-checkbox-table',
     'ovh-angular-pagination-front',
     'ovh-angular-responsive-tabs',
+    'ovh-api-services',
+    'ngMessages',
     dashboard,
     guides,
     options,
@@ -52,8 +56,8 @@ angular
   .constant('SMS_ALERTS', constant.SMS_ALERTS)
   .constant('SMS_PHONEBOOKS', constant.SMS_PHONEBOOKS)
   .config(($stateProvider) => {
-    $stateProvider.state('sms', {
-      url: '/sms/:serviceName',
+    $stateProvider.state('sms.service', {
+      url: '/:serviceName',
       views: {
         '': {
           template: smsView,