+++ packages/manager/apps/layout-ovh/CHANGELOG.md
@@ -1,3 +1,36 @@
+# [2.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-layout-ovh@1.2.5...@ovh-ux/manager-layout-ovh@2.0.0) (2019-01-29)
+
+
+### Build System
+
+* **deps:** upgrade ng-ovh-http to v4.0.1-beta.0 ([b2e4388](https://github.com/ovh-ux/manager/commit/b2e4388))
+* **deps:** upgrade ng-ovh-otrs to v7.0.0-beta.1 ([9c8da48](https://github.com/ovh-ux/manager/commit/9c8da48))
+* **deps:** upgrade ng-ovh-sidebar-menu to v8.0.0-beta.0 ([460588c](https://github.com/ovh-ux/manager/commit/460588c))
+* **deps:** upgrade ng-ovh-sso-auth to v4.0.0-beta.0 ([8acac96](https://github.com/ovh-ux/manager/commit/8acac96))
+* **deps:** upgrade ng-ovh-telecom-universe-components to v2.0.1 ([3ffc516](https://github.com/ovh-ux/manager/commit/3ffc516))
+* **deps:** upgrade ng-translate-async-loader to v2.0.0 ([40e8ea7](https://github.com/ovh-ux/manager/commit/40e8ea7))
+
+
+### BREAKING CHANGES
+
+* **deps:** replace `ovh-angular-sidebar-menu` by `@ovh-ux/ng-ovh-sidebar-menu`
+* **deps:** replace `ovh-angular-otrs` by `@ovh-ux/ng-ovh-otrs`
+* **deps:** replace `@ovh-ux/translate-async-loader` by `@ovh-ux/ng-translate-async-loader`
+* **deps:** replace `@ovh-ux/telecom-universe-components` by `@ovh-ux/ng-ovh-telecom-universe-components`
+* **deps:** replace `ovh-angular-sso-auth` by `@ovh-ux/ng-ovh-sso-auth`
+* **deps:** replace `ovh-angular-http` by `@ovh-ux/ng-ovh-http`
+
+
+
+## [1.2.5](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-layout-ovh@1.2.4...@ovh-ux/manager-layout-ovh@1.2.5) (2019-01-23)
+
+
+### Bug Fixes
+
+* **layout-ovh:** update sms.dashboard state to sms.service.dashboard ([2c526e4](https://github.com/ovh-ux/manager/commit/2c526e4))
+
+
+
 ## [1.2.4](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-layout-ovh@1.2.3...@ovh-ux/manager-layout-ovh@1.2.4) (2019-01-21)